requirements.txt
fastapi~=0.110.2
uvicorn
pycryptodome
pydantic~=2.7.1
